♻️Refactor: TrendingVideo document에서 domain 디렉토리로 변경
